requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
contourpy==1.1.0
cycler==0.11.0
dbus-next==0.2.3
desktop-notifier==3.5.6
Flask==2.3.2
fonttools==4.40.0
idna==3.4
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
Pillow==9.5.0
pip-autoremove==0.10.0
pyparsing==3.0.9
pystray==0.19.4
python-benedict==0.31.0
python-dateutil==2.8.2
python-fsutil==0.10.0
python-slugify==8.0.1
python-xlib==0.33
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
text-unidecode==1.3
urllib3==2.0.3
vext==0.7.6
vext.gi==0.7.4
Werkzeug==2.3.6
